requirements.txt
django==3.0.7
djangorestframework==3.11.0
psycopg2-binary==2.8.5
sentry-sdk==0.14.4
Pillow==7.1.2
drf-yasg==1.17.1
django-mptt==0.11.0
djangorestframework-recursive==0.1.2